% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snps.R
\name{extract_sig_snps}
\alias{extract_sig_snps}
\title{Extract SNPs with P value below a specified threshold (e.g. significant SNPs)}
\usage{
extract_sig_snps(
  path_to_target_file = NULL,
  p_val_col_number = NULL,
  p_threshold = 5e-08
)
}
\arguments{
\item{path_to_target_file}{path to the target file. This contains the summary data for the trait of interest}

\item{p_val_col_number}{the column number corresponding to the P values for the SNP-trait associations}

\item{p_threshold}{Extract SNP-trait associtions with P values less than this value. Default set to 5e-8}
}
\value{
data frame
}
\description{
Exract the rows of the summary dataset of interest with P values below the specified threshold. This only works on linux/mac operating systems.
}
