% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_gwas_catalog.R
\name{find_hits_in_gwas_catalog}
\alias{find_hits_in_gwas_catalog}
\title{Are hits in the GWAS catalog?}
\usage{
find_hits_in_gwas_catalog(
  gwas_hits = NULL,
  trait = NULL,
  efo = NULL,
  efo_id = NULL
)
}
\arguments{
\item{gwas_hits}{the "GWAS hits" in the test dataset (e.g. SNP-trait associations with P<5e-8)}

\item{trait}{the trait of interest}

\item{efo}{trait of interest in the experimental factor ontology}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}
}
\value{
list
}
\description{
Identify GWAS hits in the test dataset and see if they overlap (+/- 250kb) with GWAS hits in the GWAS catalog.
}
