% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_maf.R
\name{flag_af_conflicts}
\alias{flag_af_conflicts}
\title{Flag allele frequency conflicts}
\usage{
flag_af_conflicts(target_dat = NULL)
}
\arguments{
\item{target_dat}{the dataset of interest. Data frame.}
}
\value{
list
}
\description{
Flag allele frequency conflicts through comparison of reported allele frequency to minor allele frequency in the 1000 genomes super populations.
}
