% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_gwas_catalog.R
\name{flag_gc_conflicts}
\alias{flag_gc_conflicts}
\title{Flag conflicts with the GWAS catalog}
\usage{
flag_gc_conflicts(
  dat = NULL,
  beta = "lnor",
  se = "lnor_se",
  efo = NULL,
  trait = NULL,
  efo_id = NULL,
  gwas_catalog_ancestral_group = c("European", "East Asian"),
  exclude_palindromic_snps = TRUE
)
}
\arguments{
\item{dat}{the test dataset of interest}

\item{beta}{name of the column containing the SNP effect size}

\item{se}{name of the column containing the standard error for the SNP effect size.}

\item{efo}{trait of interest in the experimental factor ontology}

\item{trait}{the trait of interest}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}

\item{gwas_catalog_ancestral_group}{restrict the comparison to these ancestral groups in the GWAS catalog. Default is set to (c("European","East Asian")}

\item{exclude_palindromic_snps}{should the function exclude palindromic SNPs? default set to TRUE. If set to FALSE, then conflicts with the GWAS catalog could reflect comparison of different reference strands.}
}
\value{
list
}
\description{
Flag conflicts with the GWAS catalog through comparison of reported effect alleles and reported effect allele frequency.
}
