% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_gwas_catalog.R
\name{flag_gc_conflicts2}
\alias{flag_gc_conflicts2}
\title{Flag conflicts with the GWAS catalog}
\usage{
flag_gc_conflicts2(gc_dat = NULL)
}
\arguments{
\item{gc_dat}{dataset generated by compare_effect_to_gwascatalog2()}
}
\value{
list
}
\description{
Flag conflicts with the GWAS catalog through comparison of reported effect alleles and reported effect allele frequency.
}
