% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{format data}
\usage{
format_data(
  dat = NULL,
  trait = NA,
  population = NA,
  ncase = NA,
  ncontrol = NA,
  rsid = NA,
  effect_allele = NA,
  other_allele = NA,
  beta = NA,
  se = NA,
  lnor = NA,
  lnor_se = NA,
  eaf = NA,
  p = NA,
  or = NA,
  or_lci = NA,
  or_uci = NA,
  chr = NA,
  pos = NA,
  z_score = NA,
  drop_duplicate_rsids = TRUE
)
}
\arguments{
\item{dat}{the dataset to be formatted}

\item{trait}{the name of the trait.}

\item{population}{describe the population ancestry of the dataset}

\item{ncase}{number of cases or name of the column specifying the number of cases}

\item{ncontrol}{number of controls or name of the column specifying the number of controls. If your summary data was generated in a linear model of a continuous trait, use ncontrol to indicate the total sample size.}

\item{rsid}{name of the column containing the rs number or identifiers for the genetic variants}

\item{effect_allele}{name of the effect allele column}

\item{other_allele}{name of the non-effect allele column}

\item{beta}{name of the column containing the SNP effect sizes. Use this argument if your summary data was generated in a linear model of a continuous trait.}

\item{se}{standard error for the beta. Use this argument if your summary data was generated in a linear model of a continuous trait.}

\item{lnor}{name of the column containing the log odds ratio. If missing, tries to infer it from the odds ratio}

\item{lnor_se}{name of the column containing the standard error for the log odds ratio. If missing, tries to infer it from 95\% confidence intervals or pvalues}

\item{eaf}{name of the effect allele frequency column}

\item{p}{name of the pvalue columne}

\item{or}{name of column containing the odds ratio}

\item{or_lci}{name of column containing the lower 95\% confidence interval for the odds ratio}

\item{or_uci}{name of column containing the upper 95\% confidence interval for the odds ratio}

\item{chr}{name of the column containing the chromosome number for each genetic variant}

\item{pos}{genomic position for the genetic variant in base pairs}

\item{z_score}{effect size estimate divided by its standard error}

\item{drop_duplicate_rsids}{drop duplicate rsids? logical. default TRUE. duplicate rsids may for example correspond to triallelic SNPs.}
}
\value{
data frame
}
\description{
Get the trait summary data ready for the QC checks.
}
