% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{get_efo}
\alias{get_efo}
\title{get efo}
\usage{
get_efo(trait = NULL)
}
\arguments{
\item{trait}{the trait of interest}
}
\value{
list
}
\description{
Retrieve the experimental factor ontology (EFO) for some trait of interest. EFOs are retrieved from ZOOMA https://www.ebi.ac.uk/spot/zooma/
}
