% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_maf.R
\name{infer_ancestry}
\alias{infer_ancestry}
\title{Infer ancestry}
\usage{
infer_ancestry(target_dat = NULL)
}
\arguments{
\item{target_dat}{the dataset of interest. Data frame.}
}
\value{
list
}
\description{
Infer possible ancestry through comparison of allele frequency amongst test dataset and 1000 genomes super populations. Returns list of Pearson correlation coefficients.
}
