% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plots.R
\name{make_cow_plot2}
\alias{make_cow_plot2}
\title{Make cow plot}
\usage{
make_cow_plot2(
  Plot_list = NULL,
  out_file = NULL,
  return_plot = FALSE,
  width = 1000,
  height = 1000,
  Title = "",
  Xlab = "",
  Ylab = "",
  Title_size = 0,
  Title_axis_size = 10,
  bycols = TRUE,
  ncol = 2,
  nrow = 4
)
}
\arguments{
\item{Plot_list}{list of plots to combine}

\item{out_file}{filepath to save the plot}

\item{return_plot}{logical argument. If TRUE, plot is returned and is not save to out_file}

\item{width}{width of plot}

\item{height}{height of plot}

\item{Title}{plot title}

\item{Xlab}{label for X axis}

\item{Ylab}{label for Y axis}

\item{Title_size}{size of title}

\item{Title_axis_size}{size of x axis title}

\item{bycols}{logical argument. If true, plots by column}

\item{ncol}{number of columns}

\item{nrow}{number of rows}
}
\value{
plot
}
\description{
Combine all plots into a single plot using the cowplot package
}
