% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_gwas_catalog.R
\name{make_plot_gwas_catalog}
\alias{make_plot_gwas_catalog}
\title{Compare the target study to the GWAS catalog}
\usage{
make_plot_gwas_catalog(
  dat = NULL,
  plot_type = "plot_zscores",
  efo_id = NULL,
  efo = NULL,
  trait = NULL,
  gwas_catalog_ancestral_group = NULL,
  legend = TRUE,
  Title = Title,
  Title_size_subplot = Title_size_subplot,
  Ylab = Ylab,
  Xlab = Xlab,
  Title_xaxis_size = 0
)
}
\arguments{
\item{dat}{the target dataset of interest}

\item{plot_type}{compare Z scores or effect allele frequency? For comparison of Z scores set plot_type to "plot_zscores". For comparison of effect allele frequency set to "plot_eaf". Default is set to "plot_zscores"}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}

\item{efo}{trait of interest in the experimental factor ontology}

\item{trait}{the trait of interest}

\item{gwas_catalog_ancestral_group}{restrict the comparison to these ancestral groups in the GWAS catalog. Default is null and comparisons are made to any study in the GWAS catalog regardless of ancestral background. If set to (c("European","East Asian") comparisons will be restricted to GWAS catalog studies of European or East Asian ancestry}

\item{legend}{include legend in plot. Default TRUE}

\item{Title}{plot title}

\item{Title_size_subplot}{size of title}

\item{Ylab}{label for Y axis}

\item{Xlab}{label for X axis}

\item{Title_xaxis_size}{size of x axis title}
}
\value{
plot
}
\description{
Make a plot comparing signed Z scores, or effect allele frequency, between the target dataset and the GWAS catalog
}
