% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_predlnor.R
\name{make_plot_pred_effect}
\alias{make_plot_pred_effect}
\title{Predicted versus reported effect sizes}
\usage{
make_plot_pred_effect(
  dat = NULL,
  Xlab = "Reported effect size",
  Ylab = "Predicted effect size",
  subtitle = "",
  maf_filter = FALSE,
  bias = FALSE,
  Title_size = 12,
  Title = "Predicted versus reported effect size",
  Title_xaxis_size = 12,
  legend = TRUE,
  standard_errors = FALSE,
  pred_beta = "lnor_pred",
  pred_beta_se = "lnor_se_pred",
  beta = "lnor",
  se = "lnor_se",
  sd_est = "sd_est",
  exclude_1000G_MAF_refdat = TRUE
)
}
\arguments{
\item{dat}{the target dataset of interest}

\item{Xlab}{label for X axis}

\item{Ylab}{label for Y axis}

\item{subtitle}{subtitle}

\item{maf_filter}{minor allele frequency threshold. If not NULL, genetic variants with a minor allele frequency below this threshold are excluded}

\item{bias}{logical argument. If TRUE, plots the \% deviation of the predicted from the reported effect size on the Y axis against the reported effect size on the X axis.}

\item{Title_size}{size of title}

\item{Title}{plot title}

\item{Title_xaxis_size}{size of x axis title}

\item{legend}{logical argument. If true, includes figure legend in plot}

\item{standard_errors}{logical argument. If TRUE, plots the predicted versus the reported standard errors for the effect sizes}

\item{pred_beta}{name of column containing the predicted effect size}

\item{pred_beta_se}{name of column containing the standard error for the predicted effect size}

\item{beta}{name of column containing the reported effect size}

\item{se}{name of column containing the standard error for the reported effect size}

\item{sd_est}{the standard deviation of the phenotypic mean. Can either be a numeric vector of length 1 or name of the column in dat containing the standard deviation value (in which case should be constant across SNPs). Only applicable for continuous traits. If not supplied by the user, the standard deviation is approximated using sd_est, estimated by the predict_beta_sd() function. The sd_est is then used to standardise the reported effect size. If the reported effect size is already standardised (ie is in SD units) then sd_est should be set to NULL}

\item{exclude_1000G_MAF_refdat}{exclude rsids from the 1000 genome MAF reference dataset.}
}
\value{
plot
}
\description{
Make a plot comparing the predicted effect sizes to the reported effect sizes.
}
