% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snplist.R
\name{make_snplist}
\alias{make_snplist}
\title{make a SNP list}
\usage{
make_snplist(trait = NULL, efo_id = NULL, efo = NULL)
}
\arguments{
\item{trait}{the trait of interest}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}

\item{efo}{trait of intersest in the experimental factor ontology}
}
\value{
character vector
}
\description{
Identify SNPs associated with your trait of interest in the GWAS catalog
}
\examples{
snplist<-make_snplist(efo_id="EFO_0004541") 
}
