% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_predlnor.R
\name{predict_beta_sd}
\alias{predict_beta_sd}
\title{Predicted standardised beta}
\usage{
predict_beta_sd(
  dat = NULL,
  beta = "beta",
  se = "se",
  eaf = "eaf",
  sample_size = "ncontrol",
  pval = "p"
)
}
\arguments{
\item{dat}{the outcome dataset of interest}

\item{beta}{the effect size column}

\item{se}{the standard error column}

\item{eaf}{the effect allele frequency column}

\item{sample_size}{the sample size column}

\item{pval}{name of the p value column}
}
\value{
data frame with predicted standardised beta, r2 and F stat statistics and estimated standard deviation
}
\description{
Predict the standardised beta using sample sise, Z score and minor allele frequency. Returns the predicted standardised beta, proportion of phenotypic variance explained by the SNP (r2) and F statistic for each SNP
}
