% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_predlnor.R
\name{predict_lnor_sh}
\alias{predict_lnor_sh}
\title{Predicted log odds ratio}
\usage{
predict_lnor_sh(dat = NULL)
}
\arguments{
\item{dat}{the outcome dataset of interest}
}
\value{
data frame
}
\description{
Predict the log odds ratio using information on the number of cases and controls, Z scores and minor allele frequency
}
