% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_plot.R
\name{zz_plot}
\alias{zz_plot}
\title{ZZ plot}
\usage{
zz_plot(
  dat = NULL,
  Title_size = 12,
  Title = "ZZ plot",
  Ylab = "Z score inferred from p value",
  Xlab = "Z score inferred from effect size and standard error",
  Title_xaxis_size = 12,
  beta = "lnor",
  se = "lnor_se"
)
}
\arguments{
\item{dat}{the target dataset of interest}

\item{Title_size}{size of title}

\item{Title}{plot title}

\item{Ylab}{label for Y axis}

\item{Xlab}{label for X axis}

\item{Title_xaxis_size}{size of x axis title}

\item{beta}{the name of the column containing the SNP effect size}

\item{se}{the name of the column containing the standard error for the SNP effect size}
}
\value{
plot
}
\description{
Calculate Z scores from the reported P values (Zp) and the reported log odds ratios (Zlnor). Construct a scatter plot of Zp and Zlnor
}
