% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_gwas_catalog.R
\name{compare_effect_to_gwascatalog2}
\alias{compare_effect_to_gwascatalog2}
\title{Compare the genetic effect sizes in the test dataset to the GWAS catalog}
\usage{
compare_effect_to_gwascatalog2(
  dat = NULL,
  efo = NULL,
  efo_id = NULL,
  trait = NULL,
  gwas_catalog_ancestral_group = c("European", "East Asian"),
  exclude_palindromic_snps = TRUE,
  map_association_to_study = TRUE,
  beta = "beta",
  se = "se"
)
}
\arguments{
\item{dat}{the test dataset of interest}

\item{efo}{trait of interest in the experimental factor ontology}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}

\item{trait}{the trait of interest}

\item{gwas_catalog_ancestral_group}{restrict the comparison to these ancestral groups in the GWAS catalog. Default is set to (c("European","East Asian")}

\item{exclude_palindromic_snps}{should the function exclude palindromic SNPs? default set to TRUE. If set to FALSE, then conflicts with the GWAS catalog could reflect comparison of different reference strands.}

\item{map_association_to_study}{map associations to study in GWAS catalog. This supports matching of results on PMID and study ancestry, which increases accuracy of comparisons, but is slow when there are large numbers of associations. Default = TRUE}

\item{beta}{name of the column containing the SNP effect size}

\item{se}{name of the column containing the standard error for the SNP effect size.}
}
\value{
dataframe
}
\description{
Compare the direction of effects and effect allele frequency between the test dataset and the GWAS catalog, in order to identify effect allele meta data errors
}
