% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_snps.R
\name{extract_snps}
\alias{extract_snps}
\title{Extract SNPs}
\usage{
extract_snps(
  snplist = NULL,
  path_to_target_file = NULL,
  exact_match = TRUE,
  path_to_target_file_sep = "\\t",
  Test.gz = FALSE,
  fill = FALSE,
  Comment = "#",
  Head = TRUE,
  get_sig_snps = FALSE,
  p_val_col_number = NULL,
  p_threshold = 5e-08
)
}
\arguments{
\item{snplist}{a list of rsids of interest, either a character vector or path_to_target_file with the list of
rsids}

\item{path_to_target_file}{path to the target file This contains the summary data for the trait of #' interest}

\item{exact_match}{search for exact matches. Default TRUE}

\item{path_to_target_file_sep}{column/field separator. Default assumes that data is tab separated}

\item{Test.gz}{is the target data a gz file? Default set to FALSE}

\item{fill}{argument from read.table. logical. If ‘TRUE’ then in case the rows have unequal length, blank fields are implicitly added.  Default is FALSE}

\item{Comment}{comment to pass to comment.char in read.table. default = "#"}

\item{Head}{Does the file have a header ? Default set to TRUE}

\item{get_sig_snps}{also extract the top hits from the target file, not just the SNPs specified in snplist. logic TRUE or FALSE. Default set to FALSE}

\item{p_val_col_number}{the column number corresponding to the P values for the SNP-trait associations}

\item{p_threshold}{Extract SNP-trait associtions with P values less than this value. Default set to 5e-8}
}
\value{
data frame
}
\description{
Exract the summary data for the rsids of interest from a target study. This only works on linux/
mac operating systems. Will not work on Windows.
}
