% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{format data}
\usage{
format_data(
  dat = NULL,
  outcome = NA,
  population = NA,
  pmid = NA,
  study = NA,
  ncase = NA,
  ncontrol = NA,
  UKbiobank = NA,
  rsid = NA,
  effect_allele = NA,
  other_allele = NA,
  lnor = NA,
  se = NA,
  eaf = NA,
  p = NA,
  info1 = NA,
  info2 = NA,
  info3 = NA,
  info4 = NA,
  HWEp = NA,
  phet = NA,
  I2 = NA,
  Q = NA,
  Direction = NA,
  effect_allele_confirmed = FALSE,
  or = NA,
  lci = NA,
  uci = NA,
  chr = NA,
  pos = NA,
  ID = NULL,
  test_statistic = NA,
  all_summary_stats = FALSE,
  open_gwas = FALSE,
  efo = NA,
  efo_id = NA
)
}
\arguments{
\item{dat}{the the dataset to be formatted}

\item{outcome}{the name of the outcome trait}

\item{population}{describe the ancestry of the dataset}

\item{pmid}{pubmed identifier for a publication corresponding to the dataset}

\item{study}{name of the study}

\item{ncase}{number of cases or name of the column specifying the number of cases}

\item{ncontrol}{number of controls or name of the column specifying the number of controls}

\item{UKbiobank}{was the dataset generated in UK Biobank or in a dataset that overlaps with UK Biobank? default set to FALSE}

\item{rsid}{name of the column containing the rs number or identifiers for the genetic variants}

\item{effect_allele}{name of the effect allele column}

\item{other_allele}{name of the non-effect allele column}

\item{lnor}{name of the column containing the log odds ratio. If missing, tries to infer it from the odds ratio}

\item{se}{name of the column containing the standard error for the log odds ratio. If missing, tries to infer it from 95\% confidence intervals or pvalues}

\item{eaf}{name of the effect allele frequency column}

\item{p}{name of the pvalue columne}

\item{info1}{name of the column containing metrics of imputation quality, such as info or r2 scores}

\item{info2}{name of second column containing metrics of imputation quality, such as info or r2 scores}

\item{info3}{name of third column containing metrics of imputation quality, such as info or r2 scores name of the third info score column}

\item{info4}{name of fourth column containing metrics of imputation quality, such as info or r2 scores}

\item{HWEp}{name of the column containing pvalues for tests of Hardy-Weinberg equilibrium}

\item{phet}{name of the column containing pvalues for tests of between study heterogeneity}

\item{I2}{name of the column containing I2 metric of between study heterogeneity}

\item{Q}{name of the column containing Cochran Q test for between study heterogeneity}

\item{Direction}{name of the column specifying the direction of effect in each individual study within a meta-analysis of multiple studies}

\item{effect_allele_confirmed}{have you confirmed the identity of the effect allele column? default set to FALSE}

\item{or}{name of column containing the odds ratio}

\item{lci}{name of column containing the lower 95\% confidence interval for the odds ratio}

\item{uci}{name of column containing the upper 95\% confidence interval for the odds ratio}

\item{chr}{name of the column containing the chromosome number for each genetic variant}

\item{pos}{genomic position for the genetic variant in base pairs}

\item{ID}{identifier for the dataset}

\item{test_statistic}{log odds ratio divided by standard error of log odds ratio}

\item{all_summary_stats}{do you have the full set of summary data or only a subset of the summary data? Default set to FALSE}

\item{open_gwas}{were the summary data downloaded from Open GWAS https://gwas.mrcieu.ac.uk/ ? Default set to FALSE}

\item{efo}{outcome trait of interest in the experimental factor ontology.}

\item{efo_id}{ID for outcome trait of interest in the experimental factor ontology.}
}
\value{
data frame
}
\description{
Get the outcome summary data ready for the QC checks. The function assumes that the summary data was generated in a logistic regression model of a binary phenotype, e.g. cases and controls
}
\examples{
data(glioma_test_dat.RData)
dat<-format_data(dat=glioma_test_dat,outcome="Glioma",population="European",pmid=22886559,study="GliomaScan",ncase="cases",ncontrol="controls",UKbiobank=FALSE,rsid="Locus",effect_allele="Allele1",other_allele="Allele2",or="OR",lci="OR_95._CI_l",uci="OR_95._CI_u",eaf="eaf.controls",p="p",effect_allele_confirmed=TRUE,all_summary_stats=TRUE,ID=67,efo="glioma")   
}
