% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_gwas_catalog.R
\name{make_plot_gwas_catalog}
\alias{make_plot_gwas_catalog}
\title{Plot comparing the test study to the GWAS catalog}
\usage{
make_plot_gwas_catalog(
  dat = NULL,
  plot_type = "plot_zscores",
  efo_id = NULL,
  efo = NULL,
  trait = NULL,
  gwas_catalog_ancestral_group = c("European", "East Asian"),
  legend = TRUE,
  Title = "Comparison of Z scores between test dataset & GWAS catalog",
  Ylab = "Z score in test dataset",
  Xlab = "Z score in GWAS catalog",
  force_all_trait_study_hits = FALSE,
  exclude_palindromic_snps = TRUE,
  beta = "beta",
  se = "se",
  distance_threshold = 25000,
  return_dat = FALSE,
  map_association_to_study = TRUE,
  gwas_catalog = NULL,
  nocolour = FALSE
)
}
\arguments{
\item{dat}{the test dataset of interest}

\item{plot_type}{compare Z scores or effect allele frequency? For comparison of Z scores set plot_type to "plot_zscores". For comparison of effect allele frequency set to "plot_eaf". Default is set to "plot_zscores"}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}

\item{efo}{trait of interest in the experimental factor ontology}

\item{trait}{the trait of interest}

\item{gwas_catalog_ancestral_group}{restrict the comparison to these ancestral groups in the GWAS catalog. Default is set to (c("European","East Asian")}

\item{legend}{include legend in plot. Default TRUE}

\item{Title}{plot title}

\item{Ylab}{label for Y axis}

\item{Xlab}{label for X axis}

\item{force_all_trait_study_hits}{force the plot to include GWAS hits from the outcome study if they are not in the GWAS catalog? This should be set to TRUE only if dat is restricted to GWAS hits for the trait of interest. This is useful for visualising whether the outcome/trait study has an unusually larger number of GWAS hits, which could, in turn, indicate that the summary statistics have not been adequately cleaned.}

\item{exclude_palindromic_snps}{should the function exclude palindromic SNPs? default set to TRUE. If set to FALSE, then conflicts with the GWAS catalog could reflect comparison of different reference strands.}

\item{beta}{name of the column containing the SNP effect size}

\item{se}{name of the column containing the standard error for the SNP effect size.}

\item{distance_threshold}{distance threshold for deciding if the GWAS hit in the test dataset is present in the GWAS catalog. For example, a distance_threshold of 25000 means that the GWAS hit in the test dataset must be within 25000 base pairs of a GWAS catalog association, otherwise it is reported as missing from the GWAS catalog.}

\item{return_dat}{if TRUE, the dataset used to generate the plot is returned to the user and no plot is made.}

\item{map_association_to_study}{map associations to study in GWAS catalog. This supports matching of results on PMID and study ancestry, which increases accuracy of comparisons, but is slow when there are large numbers of associations. Default = TRUE}

\item{gwas_catalog}{user supplied data frame containing results from the GWAS catalog for the trait of interest. If set to NULL then the function will retrieve results from the GWAS catalog.}

\item{nocolour}{if TRUE, effect size conflicts are illustrated using shapes rather than colours. Default FALSE}
}
\value{
plot
}
\description{
Make a plot comparing signed Z scores, or effect allele frequency, between the test dataset and the GWAS catalog, in order to identify effect allele meta data errors
}
