% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_predlnor.R
\name{make_plot_predlnor}
\alias{make_plot_predlnor}
\title{Predicted log odds ratio}
\usage{
make_plot_predlnor(
  dat = NULL,
  Xlab = "",
  Ylab = "",
  linear_regression = TRUE,
  subtitle = "",
  maf_filter = FALSE,
  bias = FALSE,
  Title_size = 0,
  Title = NULL,
  Title_xaxis_size = 10,
  legend = TRUE,
  standard_errors = FALSE
)
}
\arguments{
\item{dat}{the target dataset of interest}

\item{Xlab}{label for X axis}

\item{Ylab}{label for Y axis}

\item{linear_regression}{logical argument. If TRUE the predicted log odds ratio is regressed on the reported log odds ratio. The slope and intercept are extracted from this model and shown alongside the plot}

\item{subtitle}{subtitle}

\item{maf_filter}{minor allele frequency threshold. If not NULL, genetic variants with a minor allele frequency below this threshold are excluded}

\item{bias}{logical argument. If TRUE, plots the \% deviation of the predicted from the reported log odds ratio on the Y axis against the reported log odds ratio on the X axis.}

\item{Title_size}{size of title}

\item{Title}{plot title}

\item{Title_xaxis_size}{size of x axis title}

\item{legend}{logical argument. If true, includes figure legend in plot}

\item{standard_errors}{logical argument. If TRUE, plots the predicted and  reported standard errors of the log odds ratio}
}
\value{
plot
}
\description{
make a plot comparing reported and predicted log odds ratios
}
