% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snplist.R
\name{make_snplist}
\alias{make_snplist}
\title{make a SNP list}
\usage{
make_snplist(
  trait = NULL,
  efo_id = NULL,
  efo = NULL,
  ref1000G_superpops = TRUE,
  snplist_user = NULL
)
}
\arguments{
\item{trait}{the trait of interest}

\item{efo_id}{ID for trait of interest in the experimental factor ontology}

\item{efo}{trait of interest in the experimental factor ontology}

\item{ref1000G_superpops}{include reference SNPs from 1000 genomes super populations. Default=TRUE}

\item{snplist_user}{character vector of user specified rsids.}
}
\value{
character vector
}
\description{
Create a list of rsids corresponding to "top hits" in the GWAS catalog, the 1000 genomes super popualtions and SNPs of specific interest to the user (e.g. genetic instruments/proxies for the exposure of interest).
}
\examples{
snplist<-make_snplist(efo_id="EFO_0006859") 
}
