% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_plot.R
\name{zz_plot}
\alias{zz_plot}
\title{ZZ plot}
\usage{
zz_plot(
  dat = NULL,
  Title = "ZZ plot",
  Ylab = "Z score inferred from p value",
  Xlab = "Z score inferred from effect size and standard error",
  beta = "lnor",
  se = "lnor_se",
  exclude_1000G_MAF_refdat = TRUE
)
}
\arguments{
\item{dat}{the target dataset of interest}

\item{Title}{plot title}

\item{Ylab}{label for Y axis}

\item{Xlab}{label for X axis}

\item{beta}{the name of the column containing the SNP effect size}

\item{se}{the name of the column containing the standard error for the SNP effect size}

\item{exclude_1000G_MAF_refdat}{exclude rsids from the 1000 genome MAF reference dataset.}
}
\value{
plot
}
\description{
Calculate Z scores from the reported P values (Zp) and the reported log odds ratios (Zlnor). Construct a scatter plot of Zp and Zlnor
}
