% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_plot.R
\name{zz_plot}
\alias{zz_plot}
\title{ZZ plot}
\usage{
zz_plot(
  dat = NULL,
  Title_size = 0,
  Title = "",
  Ylab = "",
  Xlab = "",
  Title_xaxis_size = 0
)
}
\arguments{
\item{dat}{the target dataset of interest}

\item{Title_size}{size of title}

\item{Title}{plot title}

\item{Ylab}{label for Y axis}

\item{Xlab}{label for X axis}

\item{Title_xaxis_size}{size of x axis title}
}
\value{
plot
}
\description{
Calculate Z scores from the reported P values (Zp) and the reported log odds ratios (Zlnor). Construct a scatter plot of Zp and Zlnor
}
